+++ src/worker/views/draftTeamHistory.ts
@@ -1,17 +1,29 @@
 import { PLAYER } from "../../common";
 import { idb } from "../db";
 import { g } from "../util";
-import type { ViewInput } from "../../common/types";
+import type {
+	ViewInput,
+	MinimalPlayerRatings,
+	Player,
+} from "../../common/types";
 
 const updateDraftTeamHistory = async (
 	inputs: ViewInput<"draftTeamHistory">,
 ) => {
+	let filter;
+	if (inputs.tid >= 0) {
+		filter = (p: Player<MinimalPlayerRatings>) => p.draft.tid === inputs.tid;
+	} else {
+		filter = (p: Player<MinimalPlayerRatings>) =>
+			p.draft.tid === g.get("userTid", p.draft.year + 1);
+	}
+
 	const stats =
 		process.env.SPORT === "basketball"
 			? ["gp", "min", "pts", "trb", "ast", "per", "ewa"]
 			: ["gp", "keyStats", "av"];
 	const playersAll2 = await idb.getCopies.players({
-		filter: p => p.draft.tid === inputs.tid,
+		filter,
 	});
 	const playersAll = await idb.getCopies.playersPlus(playersAll2, {
 		attrs: ["tid", "abbrev", "draft", "pid", "name", "age", "hof", "watch"],